Move files out of the .zip file
Having the files in the .zip file defeats much of the benefits of a Git repository. It causes the diff for any changes to be meaningless. It makes it more difficult for people to make contributions to the project. It will cause the size of the repository to grow rapidly with each update to the .zip file. It's also completely pointless because GitHub automatically generates a zip file for you via Clone or download > Download ZIP, as well as on the Releases page.